Move LocalCommand into its own module
