+++ src/lib.rs
@@ -1,7 +1,6 @@
 use bevy::prelude::*;
 use bevy::tasks::IoTaskPool;
 use bevy::tasks::Task;
-use std::ffi::OsStr;
 use std::io;
 use std::io::prelude::*;
 use std::io::BufReader;
@@ -12,6 +11,9 @@ use std::process::Command;
 use std::process::Stdio;
 use std::sync::Arc;
 use std::sync::Mutex;
+
+mod local_command;
+pub use local_command::LocalCommand;
 
 /// The ID of a process.
 pub type Pid = u32;
@@ -42,70 +44,6 @@ pub struct ProcessCompleted {
 /// The lines written to the standard output by a given process.
 #[derive(Debug, Default, Clone)]
 struct ProcessOutputBuffer(Arc<Mutex<Vec<String>>>);
-
-#[derive(Debug, Component)]
-pub struct LocalCommand {
-    pub command: Command,
-}
-
-impl LocalCommand {
-    pub fn new<S>(program: S) -> Self
-    where
-        S: AsRef<OsStr>,
-    {
-        Self {
-            command: Command::new(program),
-        }
-    }
-
-    pub fn arg<S: AsRef<OsStr>>(mut self, arg: S) -> Self {
-        self.command.arg(arg);
-        self
-    }
-
-    pub fn args<I, S>(mut self, args: I) -> Self
-    where
-        I: IntoIterator<Item = S>,
-        S: AsRef<OsStr>,
-    {
-        self.command.args(args);
-        self
-    }
-
-    pub fn env<K, V>(mut self, key: K, val: V) -> Self
-    where
-        K: AsRef<OsStr>,
-        V: AsRef<OsStr>,
-    {
-        self.command.env(key, val);
-        self
-    }
-
-    pub fn envs<I, K, V>(mut self, vars: I) -> Self
-    where
-        I: IntoIterator<Item = (K, V)>,
-        K: AsRef<OsStr>,
-        V: AsRef<OsStr>,
-    {
-        self.command.envs(vars);
-        self
-    }
-
-    pub fn env_remove<K: AsRef<OsStr>>(mut self, key: K) -> Self {
-        self.command.env_remove(key);
-        self
-    }
-
-    pub fn env_clear(mut self) -> Self {
-        self.command.env_clear();
-        self
-    }
-
-    pub fn current_dir<P: AsRef<std::path::Path>>(mut self, dir: P) -> Self {
-        self.command.current_dir(dir);
-        self
-    }
-}
 
 #[derive(Debug, Component)]
 pub struct Process {
